chore: refactor BASE_URL imports to correct path and remove unused config files
