--- src/.vscode/tasks.json
@@ -1,15 +1,0 @@
-{
-	"version": "2.0.0",
-	"tasks": [
-		{
-			"type": "npm",
-			"script": "build",
-			"group": "build",
-			"problemMatcher": [],
-			"label": "npm: build",
-			"presentation": {
-				"reveal": "silent"
-			}
-		}
-	]
-}
