+++ src/app/analytics/page.tsx
@@ -13,7 +13,7 @@ import Loader from '../../components/ui/loader/Loader';
 import Toast from '../../components/ui/Toast';
 import axios from 'axios';
 import {useRouter} from 'next/navigation';
-import { BASE_URL } from '../../config';
+import { BASE_URL } from '../../../config';
 import Link from 'next/link';
