+++ src/app/page.tsx
@@ -11,7 +11,7 @@ import LatestEvent from "./components/home/LatestEvent";
 import EventCalendar from '@/components/Calendar/EventCalendar';
 // import ChatBot from "../components/Chatbot/chatbot";
 import ServerDown from "./503/page";
-import { BASE_URL } from "../config";
+import { BASE_URL } from "../../config";
 import axios from "axios";
 
 export default function Home() {
