--- src/config.ts
@@ -1,6 +1,0 @@
-
-// export const BASE_URL = 'https://v-ticket.onrender.com/';
-export const BASE_URL = 'https://v-ticket-backend.onrender.com/';
-
-export const API_URL = 'https://vchatbot.onrender.com/api';// for production
-// export const API_URL = 'http://localhost:8080/api'; // for local development
